# docs/05-api/openapi.yaml
openapi: 3.1.0
info:
  title: Bedrijfsbeheer Dashboard API
  description: |
    RESTful API voor het Bedrijfsbeheer Dashboard systeem.

    ## Overzicht

    Deze API biedt volledige CRUD operaties voor:
    - 👥 **Users** - Gebruikersbeheer en authenticatie
    - 🔧 **Werkorders** - Werkorder management
    - 👨‍💼 **Customers** - Klantenbeheer (CRM)
    - 📦 **Inventory** - Voorraad management (Products, Materials, Services)
    - 💰 **Quotes** - Offertes
    - 🧾 **Invoices** - Facturen
    - 👔 **Employees** - Personeelsbeheer (HRM)

    ## Authenticatie

    Alle endpoints (behalve `/auth/login`) vereisen een JWT Bearer token:
    ```
    Authorization: Bearer <token>
    ```

    Token verkrijgen via POST `/auth/login`.

    ## Autorisatie

    - **Admin**: Volledige toegang (CRUD op alle resources)
    - **User**: Beperkte toegang (Read, Create, Update eigen resources)

    ## Rate Limiting

    - **Login endpoint**: 5 requests per 15 minuten
    - **Alle andere endpoints**: 100 requests per minuut

    ## Error Handling

    Alle errors volgen dit formaat:
    ```json
    {
      "error": "ResourceNotFound",
      "message": "Werkorder with id 'wo-999' not found",
      "status": 404,
      "timestamp": "2025-01-27T10:30:00Z"
    }
    ```

  version: 1.0.0
  contact:
    name: API Support
    email: support@bedrijfsbeheer.nl
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3000/api
    description: Development server
...
paths:
...
  /invoices:
    get:
      tags: [Invoices]
      summary: Get all invoices
      operationId: getInvoices
      security:
        - bearerAuth: []
      parameters:
        - name: isPaid
          in: query
          schema:
            type: boolean
        - name: customerId
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Invoice'
                  total:
                    type: integer
                  limit:
                    type: integer
                  offset:
                    type: integer

    post:
      tags: [Invoices]
      summary: Create invoice
      operationId: createInvoice
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateInvoiceInput'
      responses:
        '201':
          description: Invoice created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Invoice'
        '400':
          $ref: '#/components/responses/ValidationError'
  /invoices/{id}:
    get:
      tags: [Invoices]
      summary: Get invoice by ID
      operationId: getInvoiceById
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Invoice'
        '404':
          $ref: '#/components/responses/NotFoundError'

    put:
      tags: [Invoices]
      summary: Update invoice
      operationId: updateInvoice
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateInvoiceInput'
      responses:
        '200':
          description: Invoice updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Invoice'
        '400':
          $ref: '#/components/responses/ValidationError'
        '404':
          $ref: '#/components/responses/NotFoundError'

    delete:
      tags: [Invoices]
      summary: Delete invoice
      description: |
        **Permissions:** Admin only
      operationId: deleteInvoice
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Invoice deleted
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT Bearer token obtained from /auth/login
...
  schemas:
...
    Customer:
      type: object
      required: [id, name, email, createdAt, updatedAt]
      properties:
        id:
          type: string
          example: cust-001
        name:
          type: string
          minLength: 2
          maxLength: 200
          example: ACME Corporation
        email:
          type: string
          format: email
          example: contact@acme.com
        phone:
          type: string
          pattern: '^\+?[0-9\s\-()]{8,20}$'
          example: "+31 20 1234567"
        address:
          type: string
          maxLength: 500
          example: "Hoofdstraat 123, 1234 AB Amsterdam"
        vatNumber:
          type: string
          nullable: true
          example: "NL123456789B01"
        notes:
          type: string
          maxLength: 2000
          example: "Preferred contact via email"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    Invoice:
      type: object
      required: [id, invoiceNumber, customerId, quoteId, items, subtotal, tax, total, isPaid, createdAt, updatedAt]
      properties:
        id:
          type: string
          example: inv-001
        invoiceNumber:
          type: string
          pattern: '^INV-[0-9]{4}-[0-9]{3}$'
          example: INV-2025-001
        customerId:
          type: string
          example: cust-001
        customer:
          $ref: '#/components/schemas/Customer'
        quoteId:
          type: string
          nullable: true
          example: quote-001
        items:
          type: array
          minItems: 1
          items:
            type: object
            required: [description, quantity, unitPrice]
            properties:
              description:
                type: string
              quantity:
                type: integer
                minimum: 1
              unitPrice:
                type: number
                format: float
                minimum: 0
              total:
                type: number
                format: float
        subtotal:
          type: number
          format: float
          example: 499.98
        tax:
          type: number
          format: float
          example: 104.99
        taxRate:
          type: number
          format: float
          default: 0.21
          example: 0.21
        total:
          type: number
          format: float
          example: 604.97
        isPaid:
          type: boolean
          default: false
          example: false
        paidDate:
          type: string
          format: date-time
          nullable: true
          example: null
        paymentMethod:
          type: string
          nullable: true
          example: null
        dueDate:
          type: string
          format: date
          example: "2025-02-27"
        notes:
          type: string
          maxLength: 2000
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CreateInvoiceInput:
      type: object
      required: [customerId, items]
      properties:
        customerId:
          type: string
        quoteId:
          type: string
        items:
          type: array
          minItems: 1
          items:
            type: object
            required: [description, quantity, unitPrice]
            properties:
              description:
                type: string
              quantity:
                type: integer
              unitPrice:
                type: number
        taxRate:
          type: number
          format: float
          default: 0.21
        dueDate:
          type: string
          format: date
        notes:
          type: string
    UpdateInvoiceInput:
      type: object
      properties:
        items:
          type: array
          items:
            type: object
            properties:
              description:
                type: string
              quantity:
                type: integer
              unitPrice:
                type: number
        dueDate:
          type: string
          format: date
        notes:
          type: string

    # ==========================================
    # ERRORS
    # ==========================================
    Error:
      type: object
      required: [error, message, status]
      properties:
        error:
          type: string
          example: ValidationError
        message:
          type: string
          example: "Title must be at least 3 characters"
        status:
          type: integer
          example: 400
        timestamp:
          type: string
          format: date-time
          example: "2025-01-27T10:30:00Z"
        details:
          type: object
          description: Additional error details (field validation errors)
          additionalProperties: true
          example:
            title: "Title must be at least 3 characters"
            customerId: "Customer ID is required"
  responses:
...
    ForbiddenError:
      description: Insufficient permissions
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "Forbidden"
            message: "Only admins can delete resources"
            status: 403
    NotFoundError:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "ResourceNotFound"
            message: "Werkorder with id 'wo-999' not found"
            status: 404
    ValidationError:
      description: Invalid input data
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "ValidationError"
            message: "Invalid input data"
            status: 400
            details:
              title: "Title must be at least 3 characters"
              customerId: "Customer ID is required"
...
security:
  - bearerAuth: []